% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{antidiag}
\alias{antidiag}
\title{Matrix antidiagonals}
\arguments{
\item{a}{matrix, vector or 1D array, or missing.}

\item{ncol, }{nrow optional dimensions for the result when x is not a matrix.}
}
\description{
Extract or replace the antidiagonal of a matrix,
or construct a antidiagonal matrix.
}
