% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{are_matrices}
\alias{are_matrices}
\title{Check, is all elements are matrices}
\usage{
are_matrices(...)
}
\description{
Check, is all elements are matrices
}
\examples{
are_matrix(matrix(0, 3, 3), m(1:3 | 4:6), "trolo")
}
