% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{col_bind}
\alias{col_bind}
\title{Bind vector, single values and matrices}
\usage{
col_bind(...)
}
\description{
This function works very similar to well-known base
`cbind` function. However, there is one big difference
between these functions. If you pass a vector, each value
will be get individually.
}
\examples{
cbind(1:5)
     [,1]
[1,]    1
[2,]    2
[3,]    3
[4,]    4
[5,]    5
col_bind(1:5)
     [,1] [,2] [,3] [,4] [,5]
[1,]    1    2    3    4    5


}
