% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{crep}
\alias{crep}
\title{Repeat columns}
\usage{
crep(x, times)
}
\arguments{
\item{x}{matrix}

\item{times}{number of repetitions}
}
\value{
matrix
}
\description{
Repeat columns
}
\examples{
crep(v(1:3), 4)
}
