% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%/\%}
\alias{\%/\%}
\title{divide matrix by vector rowwise/columnwise}
\usage{
a \%/\% b
}
\arguments{
\item{a, b}{matrices/vectors}
}
\value{
Matrix/vector
}
\description{
This operator allows to do elementwise division of
two algebraic object i.e. matrices/vectors. There is one required condition
to perform such operation: at least one dimension values from both objects
must be the same
}
