% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_idx_possible}
\alias{is_idx_possible}
\title{Is idx possible in given matrix?}
\usage{
is_idx_possible(mat, idx)
}
\arguments{
\item{mat}{matrix}

\item{idx}{two-element vector}
}
\description{
Is idx possible in given matrix?
}
\examples{
is_idx_possible(matrix(0, 3, 3), c(4, 5))
is_idx_possible(matrix(0, 3, 3), c(3, 2))
}
