% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{neighbour_idx}
\alias{neighbour_idx}
\title{Get all indices in neighbourhood}
\usage{
neighbour_idx(mat, idx, mask = NULL)
}
\arguments{
\item{mat}{matrix or data.frame}

\item{idx}{two-element vector}

\item{mask}{logical matrix; optional}
}
\description{
Get list of indices, that are neighbour for the given one
}
\examples{
mat <- matrix(0, 3, 3)
neighbour_idx(mat, c(1, 2))
# With mask
mat <- matrix(0, 3, 4)
mask <- m(F, F, T, T | F, F, F, F | T, T, F, T)
neighbour_idx(mat, c(1, 2), mask = mask)
}
