% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{neighbour_idx_matrix}
\alias{neighbour_idx_matrix}
\title{Create matrix of lists, where each one contains list of neighbour field coordinates}
\usage{
neighbour_idx_matrix(mat, mask = NULL, diagonal = TRUE,
  random.select = NULL)
}
\arguments{
\item{mat}{matrix}

\item{mask}{logical matrix. Its dimensions must be identical with dimensions of mat}

\item{diagonal}{logical. get diagonal neighbours}
}
\description{
Create matrix of lists, where each one contains list of neighbour field coordinates
}
