% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{ones}
\alias{ones}
\title{Create matrix with ones values only}
\usage{
ones(nrow, ncol)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\description{
Create matrix with ones values only
}
