% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-matrix.R
\name{rboolm}
\alias{rboolm}
\title{Create matrix of random choosen boolean values}
\usage{
rboolm(nrow, ncol, true.proba = 0.5)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{numer of columns}

\item{true.proba}{probability of true values; default: 0.5}
}
\value{
a matrix
}
\description{
Create matrix of random choosen boolean values
}
\examples{
rboolm(3, 3)
rboolm(4, 5, true.proba = 0.3)
}
