% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{repetitions}
\alias{repetitions}
\alias{crep}
\alias{rrep}
\title{Repeat columns or rows}
\usage{
crep(x, times)

rrep(x, times)
}
\arguments{
\item{x}{matrix}

\item{times}{number of repetitions}
}
\value{
matrix
}
\description{
Repeat matrix object respectively to its shape and orientation
}
\details{
crep = columnwise repeat
rrep = rowwise repeat
}
\examples{
# Columnwise repetition
crep(v(1:3), 4)
crep(t(v(1:5)), 4)
# Rowwise repetition
rrep(v(1:3), 4)
rrep(t(v(1:5)), 4)
}
