% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rrep}
\alias{rrep}
\title{Repeat columns}
\usage{
rrep(x, times)
}
\arguments{
\item{x}{matrix}

\item{times}{number of repetitions}
}
\value{
matrix
}
\description{
Repeat columns
}
\examples{
rrep(v(1:3), 4)
rrep(t(v(1:5)), 4)
}
