% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-matrix.R
\name{runif_same_dims}
\alias{runif_same_dims}
\title{Create matrix of random values with dimensions copied from an existing matrix}
\usage{
runif_same_dims(mat, min = 0, max = 1)
}
\arguments{
\item{mat}{matrix}

\item{min}{lower limit of the distribution. Must be finite.}

\item{max}{upper limit of the distribution. Must be finite.}
}
\value{
a matrix
}
\description{
Create matrix of random values with dimensions copied from an existing matrix
}
\examples{
mat <- matrix(0, 3, 3)
runif_same_dims(mat)
}
