% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_matrix}
\alias{seq_matrix}
\title{Return a sequence of pairs (value, index vector)}
\usage{
seq_matrix(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
list of two-element list (single value, two-element vector)
}
\description{
Facilitates iterating over matrix, returning a sequence of pairs,
 where the first element is a value at index (x, y) and the second one is the index (x, y)
}
\examples{
mat <- matrix(1:9, 3, 3)
seq_matrix(mat)
}
