% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_values}
\alias{set_values}
\title{Set multiple values useing one function call}
\usage{
set_values(matrix, ...)
}
\description{
This functions allows to set multiple elements of a matrix
instead of using annoying step-by-step assignment by
mat[1,2] <- 2
mat[2,3] <- 0.5
etc.
}
\examples{
mat <- matrix(0, 4, 5)
set_values(mat, c(1,1) ~ 5, c(3, 4) ~ 0.3)
}
