% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{v}
\alias{v}
\title{A shortcut to create a vector}
\usage{
v(...)
}
\arguments{
\item{...}{arbitrary number of values}
}
\value{
matrix with dims n_elements x 1
}
\description{
This function provides convenient shortcut
to create a vertical (column) vector.
}
\section{Creating matrices and vectors}{
NA
}

