% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{zeros}
\alias{zeros}
\title{Create matrix with zeros values only}
\usage{
zeros(nrow, ncol)
}
\arguments{
\item{nrow}{number of rows}

\item{ncol}{number of columns}
}
\description{
Create matrix with zeros values only
}
