% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{antidiag}
\alias{antidiag}
\title{Matrix antidiagonals}
\usage{
antidiag(x = as.numeric(c(1)), nrow = NULL, ncol = NULL)
}
\arguments{
\item{x}{matrix, vector or 1D array, or missing.}

\item{nrow}{number of rows (optional; when x is not a matrix)}

\item{ncol}{number of columns (optional; when x is not a matrix)}
}
\description{
Extract or replace the antidiagonal of a matrix,
or construct a antidiagonal matrix.
}
\examples{
antidiag(diag(3))
antidiag(7, 3, 3)
antidiag(1:5, 3, 2)
}
