% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_idx}
\alias{matrix_idx}
\title{Get available marix indices}
\usage{
matrix_idx(x, n.row = NULL, n.col = NULL, mask = NULL)
}
\arguments{
\item{x}{matrix}

\item{n.row}{number of rows; default: NULL}

\item{n.col}{number of columns; default: NULL}

\item{mask}{logical matrix; default: NULL}
}
\description{
Get available marix indices
}
\examples{
mat <- matrix(0, 3, 3)
mask <- m(T, T, F | T, F, T | F, F, T)
# All poss
matrix_idx(mat)
matrix_idx(mat, mask = mask)
matrix_idx(mask = mask)
}
