% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_values}
\alias{set_values}
\alias{sv}
\title{Set multiple values useing one function call}
\usage{
set_values(mat, ...)

sv(mat, ...)
}
\arguments{
\item{mat}{a matrix object}

\item{...}{formulae; left hand values should be two-element interger vectors and right-hand: a single-value numeric}
}
\value{
matrix
}
\description{
This functions allows to set multiple elements of a matrix
instead of using annoying step-by-step assignment by
mat[1,2] <- 2
mat[2,3] <- 0.5
etc.
}
\examples{
mat <- matrix(0, 4, 5)
set_values(mat, c(1,1) ~ 5, c(3, 4) ~ 0.3)
}
