% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{v}
\alias{v}
\title{A shortcut to create a vertical vector}
\usage{
v(...)
}
\arguments{
\item{...}{arbitrary number of values}
}
\value{
matrix with dims n_elements x 1
}
\description{
This function provides convenient shortcut
to create a vertical (column) vector.
}
\examples{
# Enumerating all the values with commas
v(1, 2, 3)
# Passing whole sequence as an argument
v(1:5)
}
